+++ src/pages/profile/checkValidityProfile.ts
@@ -23,3 +23,10 @@ export function addValidationListenersToInputProfile(
     e.preventDefault();
   });
 }
+
+
+export function addShippingAddress(){
+
+
+
+}
