+++ src/pages/registration/registrationHandler.ts
@@ -30,9 +30,11 @@ export function formRegistrationHandler(event: Event) {
   if (checkboxSettingDefaultAddressShipping) customer.defaultShippingAddress = 0;
 
   if (!checkboxSettingOneAddress) {
-    customer.billingAddresses = [0];
+    const addressBilling = addressShipping;
+    customer.addresses.push(addressBilling);
+    customer.billingAddresses = [1];
     if (checkboxSettingDefaultAddressShipping) {
-      customer.defaultBillingAddress = 0;
+      customer.defaultBillingAddress = 1;
     }
   }
   if (checkboxSettingOneAddress) {
